+++ src/components/incidence/Widget.vue
